+++ packages/core/src/conversation/ConversationService/ConversationService.ts
@@ -51,6 +51,7 @@ import {GenericMessage} from '@wireapp/protocol-messaging';
 import {
   AddUsersFailureReasons,
   AddUsersParams,
+  KeyPackageClaimUser,
   MLSCreateConversationResponse,
   SendMlsMessageParams,
   SendResult,
@@ -354,7 +355,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
     const conversation = await this.getConversation(conversationId);
 
     //We store the info when user was added (and key material was created), so we will know when to renew it
-    this.mlsService.resetKeyMaterialRenewal(groupId);
+    await this.mlsService.resetKeyMaterialRenewal(groupId);
     return {
       events: response.events,
       conversation,
@@ -379,7 +380,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
     const messageResponse = await this.mlsService.removeClientsFromConversation(groupId, fullyQualifiedClientIds);
 
     //key material gets updated after removing a user from the group, so we can reset last key update time value in the store
-    this.mlsService.resetKeyMaterialRenewal(groupId);
+    await this.mlsService.resetKeyMaterialRenewal(groupId);
 
     const conversation = await this.getConversation(conversationId);
 
@@ -414,7 +415,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
       );
 
       //We store the info when user was added (and key material was created), so we will know when to renew it
-      this.mlsService.resetKeyMaterialRenewal(groupId);
+      await this.mlsService.resetKeyMaterialRenewal(groupId);
     });
   }
 
@@ -500,6 +501,14 @@ export class ConversationService extends TypedEventEmitter<Events> {
   }
 
   /**
+   * Get a MLS 1:1-conversation with a given user.
+   * @param userId - qualified user id
+   */
+  async getMLS1to1Conversation(userId: QualifiedId) {
+    return this.apiClient.api.conversation.getMLS1to1Conversation(userId);
+  }
+
+  /**
    * Will try registering mls 1:1 conversation adding the other user.
    * If it fails and the conversation is already established, it will try joining via external commit instead.
    *
@@ -518,7 +527,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
     // Before trying to register a group, check if the group is already established o backend.
     // If remote epoch is higher than 0, it means that the group was already established.
     // It's possible that we've already received a welcome message.
-    const mlsConversation = await this.apiClient.api.conversation.getMLS1to1Conversation(otherUserId);
+    const mlsConversation = await this.getMLS1to1Conversation(otherUserId);
 
     if (mlsConversation.epoch > 0) {
       this.logger.info(
@@ -540,7 +549,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
       );
 
       await this.joinByExternalCommit(mlsConversation.qualified_id);
-      return this.apiClient.api.conversation.getMLS1to1Conversation(otherUserId);
+      return this.getMLS1to1Conversation(otherUserId);
     }
 
     // If group is not established on backend,
@@ -551,7 +560,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
       await this.mlsService.registerConversation(groupId, [otherUserId, selfUser.user], selfUser);
       this.logger.info(`Conversation (id ${mlsConversation.qualified_id.id}) established successfully.`);
 
-      return this.apiClient.api.conversation.getMLS1to1Conversation(otherUserId);
+      return this.getMLS1to1Conversation(otherUserId);
     } catch (error) {
       this.logger.info(`Could not register MLS group with id ${groupId}: `, error);
 
@@ -565,6 +574,53 @@ export class ConversationService extends TypedEventEmitter<Events> {
       return this.establishMLS1to1Conversation(groupId, selfUser, otherUserId, false);
     }
   };
+
+  /**
+   * Will try to register mls group by sending an empty commit to establish it.
+   * After group was successfully established, it will try to add other users to the group.
+   *
+   * @param groupId - id of the MLS group
+   * @param conversationId - id of the conversation
+   * @param selfUserId - id of the self user
+   * @param qualifiedUsers - list of qualified users to add to the group (should not include the self user)
+   */
+  public async tryEstablishingMLSGroup({
+    groupId,
+    conversationId,
+    selfUserId,
+    qualifiedUsers,
+  }: {
+    groupId: string;
+    conversationId: QualifiedId;
+    selfUserId: QualifiedId;
+    qualifiedUsers: QualifiedId[];
+  }): Promise<void> {
+    const wasGroupEstablishedBySelfClient = await this.mlsService.tryEstablishingMLSGroup(groupId);
+
+    if (!wasGroupEstablishedBySelfClient) {
+      this.logger.info('Group was not established by self client, skipping adding users to the group.');
+      return;
+    }
+
+    this.logger.info('Group was established by self client, adding other users to the group...');
+    const usersToAdd: KeyPackageClaimUser[] = [
+      ...qualifiedUsers,
+      {...selfUserId, skipOwnClientId: this.apiClient.validatedClientId},
+    ];
+
+    const {conversation} = await this.addUsersToMLSConversation({
+      conversationId,
+      groupId,
+      qualifiedUsers: usersToAdd,
+    });
+
+    const addedUsers = conversation.members.others;
+    if (addedUsers.length > 0) {
+      this.logger.info(`Successfully added ${addedUsers} users to the group.`);
+    } else {
+      this.logger.info('No other users were added to the group.');
+    }
+  }
 
   private async handleMLSMessageAddEvent(event: ConversationMLSMessageAddEvent): Promise<HandledEventPayload | null> {
     try {