+++ packages/core/src/storage/CoreDB.ts
@@ -18,12 +18,20 @@
  */
 
 import {DBSchema, deleteDB as idbDeleteDB, IDBPDatabase, openDB as idbOpenDb} from 'idb';
-const VERSION = 1;
+const VERSION = 3;
 
 interface CoreDBSchema extends DBSchema {
   prekeys: {
     key: string;
     value: {nbPrekeys: number; highestId: number};
+  };
+  pendingProposals: {
+    key: string;
+    value: {groupId: string; firingDate: number};
+  };
+  recurringTasks: {
+    key: string;
+    value: {key: string; firingDate: number};
   };
 }
 
@@ -35,6 +43,11 @@ export async function openDB(dbName: string): Promise<CoreDatabase> {
       switch (oldVersion) {
         case 0:
           db.createObjectStore('prekeys');
+        case 1:
+          db.deleteObjectStore('prekeys');
+          db.createObjectStore('pendingProposals');
+        case 2:
+          db.createObjectStore('recurringTasks');
       }
     },
   });